requirements.txt
asgiref==3.2.10
Django==3.1
django-ckeditor==6.0.0
django-extensions==3.0.5
django-filter==2.3.0
django-js-asset==1.2.2
django-taggit==1.3.0
Pillow==7.2.0
psycopg2==2.8.5
python-slugify==4.0.1
pytz==2020.1
sqlparse==0.3.1
text-unidecode==1.3
